build: optionally clean data on build
Merged in optionally-clean-build-data (pull request #545)

Approved-by: Axel Wettervik
Approved-by: Gaël HELLÉ

+++ scripts/build-android-development.sh
@@ -4,11 +4,15 @@
 export USE_TEST_TOKEN=1
 export IS_CORDOVA_BUILD="true"
 
-# Clean Build
-rm -rf ./www
-rm -rf ./platforms/android
+# Clean build (optional)
+if [ "$1" == "clean" ]; then
+  rm -rf ./node_modules
+  rm -rf ./plugins
+  rm -rf ./www
+  rm -rf ./platforms
+fi
 
-# Ensure packages is installed
+# Ensure packages are installed
 npm install
 
 # Package using webpack
